Layering: formalize the "continent" concept
Previously, the concept of "ECMAScript implementation" initialization was highly implicit, and it was not clear that e.g. a browser would contain multiple "ECMAScript implementations", each with their own set of job queues. We make that explicit here by defining a "continent", per previous discussions in e.g. https://esdiscuss.org/topic/event-loops-in-navigated-away-from-windows and https://esdiscuss.org/topic/holy-mixed-metaphors-batman-was-event-loops-in-navigated-away-from-windows.

Since a continent now contains a set of job queues, it is now clearer (with the help of an additional non-normative note) that when doing EnqueueJob, a host should specify which continent they are working in.